+++ lib/src/services/background_service.dart
@@ -3,9 +3,7 @@ import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:flutter_background_service/flutter_background_service.dart';
-import 'package:mawaqit/src/services/salah_background_notification/adhan_audio_service.dart';
-import 'package:mawaqit/src/services/salah_background_notification/prayer_notification_service.dart';
-import 'package:mawaqit/src/services/salah_background_notification/prayer_scheduler_service.dart';
+import 'package:mawaqit/main.dart';
 import 'package:notification_overlay/notification_overlay.dart';
 import 'package:just_audio/just_audio.dart';
 import 'package:audio_session/audio_session.dart';
@@ -19,12 +17,29 @@ import '../models/times.dart';
 
 class BackgroundService with WidgetsBindingObserver {
   static final BackgroundService _instance = BackgroundService._internal();
+  static AudioPlayer? _audioPlayer;
+  static final Set<DateTime> _scheduledTimes = {};
   static bool _isInitialized = false;
+  static Timer? _notificationTimer;
+  static bool _shouldShowNotification = false;
+
+  Duration duration = Duration();
 
   factory BackgroundService() => _instance;
 
   BackgroundService._internal() {
     WidgetsBinding.instance.addObserver(this);
+  }
+
+  static String getSalahName(int index) {
+    final names = {
+      0: S.current.fajr,
+      1: S.current.duhr,
+      2: S.current.asr,
+      3: S.current.maghrib,
+      4: S.current.isha,
+    };
+    return names[index] ?? '';
   }
 
   static Future<void> initializeService() async {
@@ -34,7 +49,7 @@ class BackgroundService with WidgetsBindingObserver {
       final service = FlutterBackgroundService();
       await _stopExistingService(service);
       await _configureAndStartService(service);
-      await PrayerNotificationService.dismissNotification();
+      await dismissExistingNotification();
       _isInitialized = true;
     } catch (e) {
       _isInitialized = false;
@@ -61,12 +76,16 @@ class BackgroundService with WidgetsBindingObserver {
   }
 
   static Future<void> setNotificationVisibility(bool shouldShow) async {
-    PrayerNotificationService.setShouldShowNotification(shouldShow);
+    _shouldShowNotification = shouldShow;
     final service = FlutterBackgroundService();
 
     if (!await _ensureServiceRunning(service)) return;
 
     service.invoke('updateNotificationVisibility', {'shouldShow': shouldShow});
+  }
+
+  static Future<void> dismissExistingNotification() async {
+    await NotificationOverlay.hideNotification();
   }
 
   static Future<void> pauseBackgroundOperations() async {
@@ -75,9 +94,10 @@ class BackgroundService with WidgetsBindingObserver {
 
     service.invoke('pauseOperations');
     await Future.wait([
-      PrayerNotificationService.dismissNotification(),
-      AdhanAudioService.stopAdhan(),
+      dismissExistingNotification(),
+      _audioPlayer?.stop() ?? Future.value(),
     ]);
+    _notificationTimer?.cancel();
   }
 
   static Future<void> resumeBackgroundOperations() async {
@@ -86,6 +106,51 @@ class BackgroundService with WidgetsBindingObserver {
     if (!await _ensureServiceRunning(service)) return;
 
     service.invoke('resumeOperations');
+  }
+
+  static String getAdhanLink(MosqueConfig? mosqueConfig, {bool useFajrAdhan = false}) {
+    String baseLink = "$kStaticFilesUrl/mp3/adhan-afassy.mp3";
+
+    if (mosqueConfig?.adhanVoice?.isNotEmpty ?? false) {
+      baseLink = "$kStaticFilesUrl/mp3/${mosqueConfig!.adhanVoice!}.mp3";
+    }
+
+    if (useFajrAdhan && !baseLink.contains('bip')) {
+      baseLink = baseLink.replaceAll('.mp3', '-fajr.mp3');
+    }
+
+    return baseLink;
+  }
+
+  static Future<void> schedulePrayerTasks(
+    Times times,
+    MosqueConfig? mosqueConfig,
+    bool isAdhanVoiceEnabled,
+    int salahIndex,
+  ) async {
+    final service = FlutterBackgroundService();
+    if (!await service.isRunning()) return;
+
+    final prayerTimes = times.dayTimesStrings(AppDateTime.now(), salahOnly: true);
+    final now = AppDateTime.now();
+
+    for (var i = 0; i < prayerTimes.length; i++) {
+      final entry = prayerTimes[i];
+      final scheduleTime = _parseScheduleTime(entry, now);
+
+      if (!_shouldSchedulePrayer(scheduleTime)) continue;
+
+      final prayerConfig = _createPrayerConfig(
+        entry,
+        scheduleTime,
+        isAdhanVoiceEnabled,
+        mosqueConfig,
+        salahIndex == 0,
+        i,
+      );
+
+      _schedulePrayerTimer(service, prayerConfig, scheduleTime);
+    }
   }
 
   static Future<bool> _ensureServiceRunning(FlutterBackgroundService service) async {
@@ -118,6 +183,67 @@ class BackgroundService with WidgetsBindingObserver {
     await service.startService();
   }
 
+  static DateTime _parseScheduleTime(String entry, DateTime now) {
+    final timeParts = entry.split(':');
+    return DateTime(
+      now.year,
+      now.month,
+      now.day,
+      int.parse(timeParts[0]),
+      int.parse(timeParts[1]),
+    );
+  }
+
+  static bool _shouldSchedulePrayer(DateTime scheduleTime) {
+    if (_scheduledTimes.contains(scheduleTime)) return false;
+    final delay = scheduleTime.difference(AppDateTime.now());
+    return !delay.isNegative;
+  }
+
+  static Map<String, dynamic> _createPrayerConfig(
+    String entry,
+    DateTime scheduleTime,
+    bool isAdhanVoiceEnabled,
+    MosqueConfig? mosqueConfig,
+    bool isFajr,
+    int index,
+  ) {
+    String adhanAsset = "";
+    bool adhanFromAssets = false;
+    bool shouldPlayAdhan = false;
+
+    if (isAdhanVoiceEnabled) {
+      shouldPlayAdhan = true;
+      final url = getAdhanLink(mosqueConfig, useFajrAdhan: isFajr);
+      if (url.contains('bip')) {
+        adhanFromAssets = true;
+        adhanAsset = R.ASSETS_VOICES_ADHAN_BIP_MP3;
+      } else {
+        adhanAsset = url;
+      }
+    }
+    final notificationFormat = S.current.prayerTimeNotification("{{salah}}", "{{time}}");
+    return {
+      'prayer': entry,
+      'time': scheduleTime.toString(),
+      'shouldPlayAdhan': shouldPlayAdhan,
+      'adhanAsset': adhanAsset,
+      'adhanFromAssets': adhanFromAssets,
+      'salahName': getSalahName(index),
+      'notificationFormat': notificationFormat
+    };
+  }
+
+  static void _schedulePrayerTimer(
+    FlutterBackgroundService service,
+    Map<String, dynamic> config,
+    DateTime scheduleTime,
+  ) {
+    final delay = scheduleTime.difference(AppDateTime.now());
+    Timer(delay, () => service.invoke('prayerTime', config));
+    _scheduledTimes.add(scheduleTime);
+  }
+
   // Background service handlers
   @pragma('vm:entry-point')
   static Future<bool> onIosBackground(ServiceInstance service) async {
@@ -130,13 +256,15 @@ class BackgroundService with WidgetsBindingObserver {
   static void onStart(ServiceInstance service) async {
     DartPluginRegistrant.ensureInitialized();
 
+    bool shouldShowNotification = false;
     bool isPaused = false;
 
-    _setupServiceListeners(service, isPaused);
+    _setupServiceListeners(service, shouldShowNotification, isPaused);
   }
 
   static void _setupServiceListeners(
     ServiceInstance service,
+    bool shouldShowNotification,
     bool isPaused,
   ) {
     if (service is AndroidServiceInstance) {
@@ -148,26 +276,32 @@ class BackgroundService with WidgetsBindingObserver {
 
     service.on('updateNotificationVisibility').listen((event) {
       if (event?['shouldShow'] != null) {
-        PrayerNotificationService.setShouldShowNotification(event!['shouldShow'] as bool);
+        shouldShowNotification = event!['shouldShow'] as bool;
       }
     });
 
     service.on('pauseOperations').listen((_) {
       isPaused = true;
-      AdhanAudioService.stopAdhan();
-      PrayerNotificationService.dismissNotification();
+      _audioPlayer?.stop();
+      dismissExistingNotification();
+      _notificationTimer?.cancel();
     });
 
     service.on('resumeOperations').listen((_) => isPaused = false);
 
     service.on('prayerTime').listen((event) async {
       if (event != null && !isPaused) {
-        await _handlePrayerTime(event);
+        await _handlePrayerTime(event, shouldShowNotification);
       }
     });
   }
 
-  static Future<void> _handlePrayerTime(Map<dynamic, dynamic> event) async {
+  static Future<void> _handlePrayerTime(
+    Map<dynamic, dynamic> event,
+    bool shouldShowNotification,
+  ) async {
+    if (!shouldShowNotification) return;
+
     final prayerName = event['prayer'] as String;
     final shouldPlayAdhan = event['shouldPlayAdhan'] as bool;
     final adhanAsset = event['adhanAsset'] as String;
@@ -175,30 +309,75 @@ class BackgroundService with WidgetsBindingObserver {
     final salahName = event['salahName'] as String;
     final notificationFormat = event['notificationFormat'] as String;
 
+    _audioPlayer = AudioPlayer();
+
     try {
+      await dismissExistingNotification();
       final notificationText = notificationFormat.replaceAll('{{salah}}', salahName).replaceAll('{{time}}', prayerName);
-
-      await NotificationOverlay.showNotification(notificationText);
-
-      if (shouldPlayAdhan) {
-        await AdhanAudioService.playAdhan(adhanAsset, adhanFromAssets);
+      try {
+        await NotificationOverlay.showNotification(notificationText);
+      } catch (e) {
+        print(e);
       }
+      if (shouldPlayAdhan) {
+        await _playPrayerAudio(adhanAsset, adhanFromAssets);
+      }
+
+      _scheduleNotificationDismissal();
     } catch (e) {
-      await PrayerNotificationService.dismissNotification();
+      await dismissExistingNotification();
     }
   }
 
-  static Future<void> schedulePrayerTasks(
-    Times times,
-    MosqueConfig? mosqueConfig,
-    bool isAdhanVoiceEnabled,
-    int salahIndex,
-  ) async {
-    await PrayerSchedulerService.schedulePrayerTasks(
-      times,
-      mosqueConfig,
-      isAdhanVoiceEnabled,
-      salahIndex,
+  static Future<void> _playPrayerAudio(String adhanAsset, bool adhanFromAssets) async {
+    final session = await _configureAudioSession();
+    await session.setActive(true);
+
+    try {
+      if (adhanFromAssets) {
+        await _audioPlayer?.setAsset(adhanAsset);
+      } else {
+        await _audioPlayer?.setUrl(adhanAsset);
+      }
+
+      await _audioPlayer?.play();
+
+      _audioPlayer?.playbackEventStream.listen((event) {
+        if (event.processingState == ProcessingState.completed) {
+          session.setActive(false);
+          dismissExistingNotification();
+        }
+      });
+    } catch (e) {
+      await session.setActive(false);
+      await dismissExistingNotification();
+    }
+  }
+
+  static Future<AudioSession> _configureAudioSession() async {
+    final session = await AudioSession.instance;
+    await session.configure(AudioSessionConfiguration(
+      avAudioSessionCategory: AVAudioSessionCategory.playback,
+      avAudioSessionCategoryOptions: AVAudioSessionCategoryOptions.mixWithOthers,
+      avAudioSessionMode: AVAudioSessionMode.defaultMode,
+      avAudioSessionRouteSharingPolicy: AVAudioSessionRouteSharingPolicy.defaultPolicy,
+      avAudioSessionSetActiveOptions: AVAudioSessionSetActiveOptions.none,
+      androidAudioAttributes: const AndroidAudioAttributes(
+        contentType: AndroidAudioContentType.music,
+        flags: AndroidAudioFlags.audibilityEnforced,
+        usage: AndroidAudioUsage.alarm,
+      ),
+      androidAudioFocusGainType: AndroidAudioFocusGainType.gain,
+      androidWillPauseWhenDucked: true,
+    ));
+    return session;
+  }
+
+  static void _scheduleNotificationDismissal() {
+    _notificationTimer?.cancel();
+    _notificationTimer = Timer(
+      const Duration(minutes: 5),
+      dismissExistingNotification,
     );
   }
 }
