--- lib/src/services/salah_background_notification/adhan_audio_service.dart
@@ -1,72 +1,0 @@
-import 'package:just_audio/just_audio.dart';
-import 'package:audio_session/audio_session.dart';
-import 'package:mawaqit/src/const/constants.dart';
-import 'package:mawaqit/src/models/mosqueConfig.dart';
-import 'package:mawaqit/src/services/salah_background_notification/prayer_notification_service.dart';
-
-class AdhanAudioService {
-  static AudioPlayer? _audioPlayer;
-
-  static String getAdhanLink(MosqueConfig? mosqueConfig, {bool useFajrAdhan = false}) {
-    String baseLink = "$kStaticFilesUrl/mp3/adhan-afassy.mp3";
-
-    if (mosqueConfig?.adhanVoice?.isNotEmpty ?? false) {
-      baseLink = "$kStaticFilesUrl/mp3/${mosqueConfig!.adhanVoice!}.mp3";
-    }
-
-    if (useFajrAdhan && !baseLink.contains('bip')) {
-      baseLink = baseLink.replaceAll('.mp3', '-fajr.mp3');
-    }
-
-    return baseLink;
-  }
-
-  static Future<void> playAdhan(String adhanAsset, bool adhanFromAssets) async {
-    _audioPlayer = AudioPlayer();
-    final session = await _configureAudioSession();
-    await session.setActive(true);
-
-    try {
-      if (adhanFromAssets) {
-        await _audioPlayer?.setAsset(adhanAsset);
-      } else {
-        await _audioPlayer?.setUrl(adhanAsset);
-      }
-
-      await _audioPlayer?.play();
-
-      _audioPlayer?.playbackEventStream.listen((event) {
-        if (event.processingState == ProcessingState.completed) {
-          session.setActive(false);
-          PrayerNotificationService.dismissNotification();
-        }
-      });
-    } catch (e) {
-      await session.setActive(false);
-      await PrayerNotificationService.dismissNotification();
-    }
-  }
-
-  static Future<void> stopAdhan() async {
-    await _audioPlayer?.stop();
-  }
-
-  static Future<AudioSession> _configureAudioSession() async {
-    final session = await AudioSession.instance;
-    await session.configure(AudioSessionConfiguration(
-      avAudioSessionCategory: AVAudioSessionCategory.playback,
-      avAudioSessionCategoryOptions: AVAudioSessionCategoryOptions.mixWithOthers,
-      avAudioSessionMode: AVAudioSessionMode.defaultMode,
-      avAudioSessionRouteSharingPolicy: AVAudioSessionRouteSharingPolicy.defaultPolicy,
-      avAudioSessionSetActiveOptions: AVAudioSessionSetActiveOptions.none,
-      androidAudioAttributes: const AndroidAudioAttributes(
-        contentType: AndroidAudioContentType.music,
-        flags: AndroidAudioFlags.audibilityEnforced,
-        usage: AndroidAudioUsage.alarm,
-      ),
-      androidAudioFocusGainType: AndroidAudioFocusGainType.gain,
-      androidWillPauseWhenDucked: true,
-    ));
-    return session;
-  }
-}
